% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_confseq.R
\name{asymptotic_confseq}
\alias{asymptotic_confseq}
\title{Asymptotic confidence sequence}
\usage{
asymptotic_confseq(
  x,
  t_opt,
  alpha = 0.05,
  var = NULL,
  LIL = FALSE,
  return_all_times = TRUE
)
}
\arguments{
\item{x}{The observed data points (a real vector).}

\item{t_opt}{The time for which the confidence sequence should be tightest.}

\item{alpha}{The significance level (a (0, 1)-valued real).}

\item{var}{The known or estimated variance of the observations, `x`.
If left `NULL`, then the empirical variance of `x` will be
taken (positive real-valued vector or `NULL`).}

\item{LIL}{Use margin with law of the iterated logarithm rate? (boolean)}

\item{return_all_times}{Should the CS be returned at
every time point? (boolean)}
}
\value{
A list containing the following vectors: \cr
\item{l}{The lower confidence sequence (a real vector).}
\item{u}{The upper confidence sequence (a real vector).}
}
\description{
Asymptotic confidence sequence
}
