% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{betting_cs}
\alias{betting_cs}
\title{Betting confidence sequence}
\usage{
betting_cs(
  x,
  alpha,
  breaks = 1000,
  running_intersection = FALSE,
  parallel = FALSE,
  convex_comb = FALSE,
  theta = 1/2,
  trunc_scale = 1/2,
  m_trunc = TRUE
)
}
\arguments{
\item{x}{A vector of data (a real vector).}

\item{alpha}{The desired type-I error level (a real in (0, 1)).}

\item{breaks}{The number of breaks to use for the CS.}

\item{running_intersection}{Whether to use the running intersection}

\item{parallel}{Whether to use parallel processing}

\item{convex_comb}{Whether to use convex combination}

\item{theta}{The convex combination parameter}

\item{trunc_scale}{The truncation scale}

\item{m_trunc}{Whether to scale truncation based on m}
}
\value{
A list with two elements, `l` and `u`, which are vectors of lower
       and upper confidence bounds, respectively.
}
\description{
Betting confidence sequence
}
