% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cumul_var}
\alias{cumul_var}
\title{Cumulative variance}
\usage{
cumul_var(x)
}
\arguments{
\item{x}{The real-valued data}
}
\value{
The sample variance of `x` at each time
}
\description{
Cumulative variance
}
