% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_cumul_miscoverage_rate}
\alias{get_cumul_miscoverage_rate}
\title{Get the cumulative empirical miscoverage rate of a confidence sequence}
\usage{
get_cumul_miscoverage_rate(
  data_generator_fn,
  conf_set_fn,
  times,
  num_repeats,
  mu = 0,
  n_cores = 1
)
}
\arguments{
\item{data_generator_fn}{A function which generates a vector of data
(() -> real)}

\item{conf_set_fn}{A function which takes a vector of data and produces a
sequence of confidence intervals or a confidence sequence
(real -> (real, real))}

\item{times}{The times for which the miscoverage should be considered.
For example, we might want to start the experiment late since
early start times can lead to poor asymptotic approximations
(a vector of positive integers)}

\item{num_repeats}{How many repeats to perform. Higher values lead to
better estimates of the miscoverage rate (positive int).}

\item{mu}{The parameter which `conf_set_fn` should be covering (a real).}

\item{n_cores}{How many cores to parallel process on (positive int).}
}
\value{
A vector of (increasing) miscoverage rates at each time
        (a (0, 1)-valued vector).
}
\description{
Get the cumulative empirical miscoverage rate of a confidence sequence
}
