% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_robust.R
\name{get_SL_fn}
\alias{get_SL_fn}
\title{Get the Super Learner function given a library of SuperLearner algorithms.}
\usage{
get_SL_fn(
  SL.library = c("SL.earth", "SL.gam", "SL.glm", "SL.glmnet", "SL.glm.interaction",
    "SL.ranger"),
  family = gaussian()
)
}
\arguments{
\item{SL.library}{Character vector containing SuperLearner algorithm names.}

\item{family}{The family for the algorithm. If estimating the propensity
score, set `family = binomial`, but if estimating the
regression function, set `family = gaussian()`.}
}
\value{
Function which takes three arguments `y`, `X`, and `newX`, the
        training outcome, training covariates, and evaluation covariates,
        respectively. This function takes `y` and `X` to train the
        algorithm and uses the `SuperLearner` library to predict the outcome
        for the covariates `newX`.
}
\description{
Function signature: (char, family ->
                     (real, data.frame, data.frame -> real))
}
