% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{howard_subexponential_cs}
\alias{howard_subexponential_cs}
\title{Howard et al. sub-exponential confidence sequence}
\usage{
howard_subexponential_cs(
  x,
  alpha,
  v_opt,
  running_intersection = FALSE,
  lower_bd = 0,
  upper_bd = 1
)
}
\arguments{
\item{x}{A vector of data in [0, 1] (a real vector).}

\item{alpha}{The desired type-I error level (a real in (0, 1)).}

\item{v_opt}{The time to optimize the CS for (a positive integer).}
}
\value{
A list with two elements, `l` and `u`, which are vectors of lower
       and upper confidence bounds, respectively.
}
\description{
Howard et al. sub-exponential confidence sequence
}
