% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logseq}
\alias{logseq}
\title{Get a sequence of logarithmically-spaced numbers}
\usage{
logseq(from, to, n)
}
\arguments{
\item{from}{The starting value of the sequence}

\item{to}{The final value of the sequence}

\item{n}{How many values of the sequence to output}
}
\value{
A sequence of logarithmically-spaced numbers from `from` to `to`.
}
\description{
Get a sequence of logarithmically-spaced numbers
}
