% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_confseq.R
\name{lyapunov_asympcs}
\alias{lyapunov_asympcs}
\title{Lyapunov-style AsympCS}
\usage{
lyapunov_asympcs(x, rho2, alpha = 0.05, var_t = NULL, return_all_times = TRUE)
}
\arguments{
\item{x}{The observed data points (a real vector).}

\item{alpha}{The significance level (a (0, 1)-valued real).}

\item{return_all_times}{Should the CS be returned at
every time point? (boolean)}

\item{rho}{The time for which the confidence sequence should be tightest.}

\item{var}{The known or estimated variance of the observations, `x`.
If left `NULL`, then the empirical variance of `x` will be
taken (positive real-valued vector or `NULL`).}
}
\value{
A list containing the following vectors: \cr
\item{l}{The lower confidence sequence (a real vector).}
\item{u}{The upper confidence sequence (a real vector).}
}
\description{
Lyapunov-style AsympCS
}
