% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{naive_confidence_intervals}
\alias{naive_confidence_intervals}
\title{Naive confidence interval}
\usage{
naive_confidence_intervals(
  x,
  alpha = 0.05,
  var = NULL,
  return_all_times = FALSE
)
}
\arguments{
\item{x}{The observed data points (a real vector).}

\item{alpha}{The significance level (a (0, 1)-valued real).}

\item{var}{The known or estimated variance of the observations, `x`.
If left `NULL`, then the empirical variance of `x` will be
taken (positive real-valued vector or `NULL`).}

\item{return_all_times}{Should the confidence sequence be returned at each
time? (boolean)}
}
\value{
A list containing the following vectors: \cr
\item{l}{The lower confidence interval (a real vector).}
\item{u}{The upper confidence interval (a real vector).}
}
\description{
Function signature: (real, real, real, boolean) -> (real, real)
}
