% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{naive_std_margin}
\alias{naive_std_margin}
\title{Confidence interval margin for the mean of unit variance Gaussians}
\usage{
naive_std_margin(t, alpha)
}
\arguments{
\item{t}{The times at which the margin will be evaluated}

\item{alpha}{The (0, 1)-valued confidence level}
}
\value{
A vector containing the margin for each $t$
}
\description{
Confidence interval margin for the mean of unit variance Gaussians
}
