% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_confseq.R
\name{plot_cs_shape}
\alias{plot_cs_shape}
\title{Plot ratio of confidence sequence to confidence interval widths}
\usage{
plot_cs_shape(t_opts, t, alpha = 0.05, log_scale = FALSE)
}
\arguments{
\item{t_opts}{The times for which to optimize the confidence sequence
(vector of positive integers)}

\item{t}{The times to plot the confidence sequence
(vector of positive integers)}

\item{alpha}{The significance level, e.g. set alpha=0.05 for 95% coverage
(real number between 0 and 1)}

\item{log_scale}{Should the plot be returned on a log-scale? (boolean)}
}
\value{
A ggplot2 plot object
}
\description{
Plot ratio of confidence sequence to confidence interval widths
}
