% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_robust.R
\name{pseudo_outcome_sequential}
\alias{pseudo_outcome_sequential}
\title{Doubly-robust variables at many timepoints for binary treatment given nuisance function estimators.}
\usage{
pseudo_outcome_sequential(
  y,
  X,
  treatment,
  regression_fn_1,
  regression_fn_0 = NULL,
  propensity_score_fn,
  train_idx = NULL,
  times = NULL,
  n_cores = 1,
  cross_fit = FALSE
)
}
\arguments{
\item{y}{The measured outcome (a real vector).}

\item{X}{Measured covariates (an nxd real matrix where `n = length(y)`)}

\item{treatment}{Whether the subject received treatment
(a boolean vector or 0/1-valued integer vector).}

\item{regression_fn_1}{A function which predicts outcomes for those who took
treatment. This function takes three arguments: `y`,
`X`, and `newX`, the training outcome, training
covariates, and evaluation covariates, respectively.
The function outputs the predicted response given the
evaluation covariates, `newX`.}

\item{regression_fn_0}{The same as `regression_fn_1` but for those who did
not receive treatment. If left NULL, this will be
set to the same function as `regression_fn_1`.}

\item{train_idx}{The indices indicating the training split for the sample
splitting algorithm. If left NULL, the training index will
be assigned randomly with probability 1/2.}

\item{times}{The times for which the doubly-robust variables should be
calculated. Can be a vector of times (an integer vector) or
a single time (integer). If left NULL, the variables will
only be computed at time n.}

\item{n_cores}{The number of cores to use for parallelization.}

\item{propensity_score}{A function which predicts the propensity score for
each subject. Similar to `regression_fn_1`, this
function takes three arguments: `y`, `X`, and `newX`,
the training treatment indicator (1 if treatment, 0
if control), the training covariates, and the
evaluation covariates. The function outputs the
predicted propensity score given the evaluation
covariates, `newX`.}
}
\value{
A list of doubly-robust variable vectors, one vector for each time
        requested in `times`.
}
\description{
Function signature: (real, data.frame, int,
                     (real, data.frame, data.frame -> real),
                     (real, data.frame, data.frame -> real),
                     (int, data.frame, data.frame -> real),
                     int, int) -> [real]
}
