% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_robust.R
\name{pseudo_outcome_variance}
\alias{pseudo_outcome_variance}
\title{Variance of the doubly robust observations}
\usage{
pseudo_outcome_variance(y, reg_1, reg_0, propensity_score, treatment)
}
\arguments{
\item{y}{The measured outcome (a real number).}

\item{reg_1}{The predicted outcome in the treatment group
(a real vector).}

\item{reg_0}{The predicted outcome in the control group
(a real vector).}

\item{propensity_score}{The predicted (or known) propensity score
(a (0, 1)-valued vector)}

\item{treatment}{Whether the subject received treatment
(a boolean vector or 0/1-valued integer vector).}
}
\value{
The variance of the doubly-robust observations
        (a vector of real numbers)
}
\description{
Function signature: (real, real, real, real, int) -> real
}
