% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{robbins_subGaussian_cs}
\alias{robbins_subGaussian_cs}
\title{Robbins' sub-Gaussian confidence sequence}
\usage{
robbins_subGaussian_cs(x, alpha, t_opt, sigma2)
}
\arguments{
\item{x}{A vector of data (a real vector).}

\item{alpha}{The desired type-I error level (a real in (0, 1)).}

\item{t_opt}{The time to optimize the CS for (a positive integer).}

\item{sigma2}{The sub-Gaussian parameter (a real greater than 0).}
}
\value{
A list with two elements, `l` and `u`, which are vectors of lower
       and upper confidence bounds, respectively.
}
\description{
Robbins' sub-Gaussian confidence sequence
}
