% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_confseq.R
\name{std_conjmix_margin}
\alias{std_conjmix_margin}
\title{Conjugate mixture standard (unit-variance) margin}
\usage{
std_conjmix_margin(t, rho2, alpha = 0.05/2)
}
\arguments{
\item{t}{The times at which to product the margin
(a positive integer vector).}

\item{rho2}{The tuning parameter (a positive real).}

\item{alpha}{The significance level (a (0, 1)-valued real).}
}
\value{
The standard conjugate mixture margins (a real vector).
}
\description{
Function signature: (int, real, real) -> real
}
