% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_confseq.R
\name{std_LIL_margin}
\alias{std_LIL_margin}
\title{LIL standard (unit-variance) margin}
\usage{
std_LIL_margin(t, alpha = 0.05/2)
}
\arguments{
\item{t}{The times at which to product the margin
(a positive integer vector).}

\item{alpha}{The significance level (a (0, 1)-valued real).}
}
\value{
The standard LIL margins (a real vector).
}
\description{
Function signature: (int, real) -> real
}
