% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_confseq.R
\name{best_rho2_exact}
\alias{best_rho2_exact}
\title{Get the best value of $\rho^2$ for a time t_opt (exact optimization)}
\usage{
best_rho2_exact(t_opt, alpha_opt = 0.05/2)
}
\arguments{
\item{t_opt}{The time for which $\rho^2$
should be optimized (a positive integer)}

\item{alpha}{The significance level (a (0, 1)-valued real).}
}
\value{
The corresponding value of $\rho^2$ (positive real)
}
\description{
Function signature: (int, real) -> real
}
