% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confseq_ate.R
\name{confseq_ate_unadjusted}
\alias{confseq_ate_unadjusted}
\title{Unadjusted estimator}
\usage{
confseq_ate_unadjusted(
  y,
  treatment,
  propensity_score,
  t_opt,
  alpha = 0.05,
  times = NULL
)
}
\description{
Unadjusted estimator
}
